Extract MetaMap to its own file
