+++ src/runtime/src/lib.rs
@@ -5,6 +5,7 @@ mod error;
 mod external;
 mod frame;
 mod logger;
+mod meta_map;
 pub mod num2;
 pub mod num4;
 pub mod value;
@@ -24,13 +25,14 @@ pub use {
     koto_bytecode::{CompilerError, Loader, LoaderError},
     koto_parser::ParserError,
     logger::{DefaultLogger, KotoLogger},
+    meta_map::{BinaryOp, MetaKey, MetaMap, UnaryOp},
     num2::Num2,
     num4::Num4,
     value::{RuntimeFunction, Value},
     value_iterator::{IntRange, ValueIterator, ValueIteratorOutput},
     value_key::{ValueKey, ValueRef},
     value_list::{ValueList, ValueVec},
-    value_map::{BinaryOp, MetaKey, UnaryOp, ValueHashMap, ValueMap},
+    value_map::{ValueHashMap, ValueMap},
     value_number::ValueNumber,
     value_string::ValueString,
     value_tuple::ValueTuple,
